janusq/chocoq/requirements.txt

annotated-types==0.7.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
appdirs==1.4.4
asttokens==2.4.1
autoray==0.6.12
cachetools==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.0
future==1.0.0
gurobipy==11.0.3
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.59.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mqt.ddsim==1.22.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.3
openqasm3==0.5.0
packaging==24.2
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.0
pyparsing==3.2.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pyzmq==26.2.0
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-ibm-runtime==0.33.2
qiskit-qasm3-import==0.5.0
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
scipy==1.14.0
semantic-version==2.10.0
six==1.16.0
stack-data==0.6.3
stevedore==5.2.0
symengine==0.13.0
sympy==1.13.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
